requirements.txt
attrs==22.1.0
exceptiongroup==1.0.0
iniconfig==1.1.1
install==1.3.5
numexpr==2.7.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tomli==2.0.1
